Don't store the PS1 scripts
Prior to this, the module creates the set and check scripts as files on
disk. This meant that the passwords and script information would be left
hanging around.

This fixes that by building the check and set scripts only when running
the exec. The exec's "command" and "onlyif" parameters now build the
templates directly rather than use the scripts on disk.

I've left the "scripts_dir" parameter around for backwards
compatibility. It doesn't do anything, but leaving it around means that
people's existing code won't break.

In a X release of this module, that deprecated parameter should be
removed.